requirements.txt
yt-dlp
pydantic_settings
google-genai